refactor: switch pnpm version

+++ pnpm/src/main.ts
@@ -11,6 +11,7 @@ import {
   type Config,
 } from '@pnpm/config'
 import { executionTimeLogger, scopeLogger } from '@pnpm/core-loggers'
+import { prependDirsToPath } from '@pnpm/env.path'
 import { filterPackagesFromDir } from '@pnpm/filter-workspace-packages'
 import { globalWarn, logger } from '@pnpm/logger'
 import { type ParsedCliArgs } from '@pnpm/parse-cli-args'
@@ -337,19 +338,26 @@ async function switchToCorrectPnpm (packageManagerFieldValue: string, config: Co
   if (pm.name !== 'pnpm' || pm.version == null || pm.version === packageManager.version) return
   const pkgName = detectIfCurrentPkgIsExecutable() ? '@pnpm/exe' : 'pnpm'
   const dir = path.join(config.pnpmHomeDir, '.tools', pkgName, pm.version)
-  fs.mkdirSync(dir, { recursive: true })
-  fs.writeFileSync(path.join(dir, 'package.json'), '{}')
-  await pnpmCmds.add(
-    {
-      ...config,
-      dir,
-      lockfileDir: dir,
-      bin: path.join(dir, 'bin'),
-    },
-    [`${pkgName}@${pm.version}`]
-  )
-  const { status } = spawn.sync(path.join(dir, 'bin/pnpm'), process.argv.slice(2), {
+  if (!fs.existsSync(dir)) {
+    fs.mkdirSync(dir, { recursive: true })
+    fs.writeFileSync(path.join(dir, 'package.json'), '{}')
+    await pnpmCmds.add(
+      {
+        ...config,
+        dir,
+        lockfileDir: dir,
+        bin: path.join(dir, 'bin'),
+      },
+      [`${pkgName}@${pm.version}`]
+    )
+  }
+  const pnpmEnv = prependDirsToPath([path.join(dir, 'bin')])
+  const { status } = spawn.sync('pnpm', process.argv.slice(2), {
     stdio: 'inherit',
+    env: {
+      ...process.env,
+      [pnpmEnv.name]: pnpmEnv.value,
+    },
   })
   process.exit(status ?? 0)
 }
